Extract forest from IFN data

Creates a forest object from Spanish Forest Inventory (IFN) data (DGCN 2005).

IFN2forest(
  IFNtreeData,
  IFNshrubData,
  ID,
  SpParams,
  IFNherbData = NULL,
  patchsize = 10000,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  verbose = TRUE
)

IFN2forestlist(
  IFNtreeData,
  IFNshrubData,
  SpParams,
  IFNherbData = NULL,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  verbose = TRUE
)

populateShrubAllometriesFromMedfuels(
  SpParams,
  coef_mapping,
  sp_params_allom,
  group_params_allom,
  species_groups
)

Arguments

IFNtreeData
A data frame with measured tree data.

IFNshrubData
A data frame with measured shrub data.

ID
A string with the ID of the plot to be extracted.

SpParams
A data frame with species parameters (see SpParamsMED).

IFNherbData
A data frame with cover and mean height of the herb layer.

patchsize
The area of the forest stand, in square meters.

setDefaults
Initializes default values for missing fields in IFN data.

filterWrongRecords
Filters wrong records (records with missing values, zero values or wrong growth forms)

verbose
A boolean flag to indicate console output.

coef_mapping
A named string vector specifying which data column should used to populate each medfate param. Elements are data base columns and names are medfate params.

sp_params_allom
A data table of species allometric coefficients (typically from package medfuels)

group_params_allom
A data table of group allometric coefficients (typically from package medfuels)

species_groups
A data table specifying raunkiaer forms for many species (typically from package medfuels)

Value

Function IFN2forest returns an object of class forest, whereas function IFN2forestlist returns a list of forest objects.

Details

IFN input data needs to be in a specific format, see package IFNread. Functions IFN2forest and IFN2forestlist call translateIFNSpeciesCodes internally.

References

DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

data(SpParamsMED)
data(exampletreedata)
data(exampleshrubdata)

# Builds an object 'forest' corresponding to one specific forest plot
f = IFN2forest(exampletreedata, exampleshrubdata,
               ID = "81065", SpParams = SpParamsMED)

#> Filtered records in tree data: 11
#> Shrub data records with unrecognized IFN species codes: 1/10 (10%)
#> 
#> 141 
#>   1 

print(f)

#> $ID
#> [1] "81065"
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#>    Species   N   DBH Height      Z50 Z95
#> 1      158 127  9.55    750 173.1933 813
#> 2      158 127  9.40    650 173.1933 813
#> 3      158  32 21.00   1600 173.1933 813
#> 4      158  14 23.30   1450 173.1933 813
#> 5      158  32 22.15   1450 173.1933 813
#> 6      158  14 23.10   1450 173.1933 813
#> 7      158  32 19.00   1350 173.1933 813
#> 8      158  14 25.40   1450 173.1933 813
#> 9      158  32 18.50   1250 173.1933 813
#> 10     158 127  8.20    750 173.1933 813
#> 11     158  32 21.25   1400 173.1933 813
#> 12     158  32 19.30   1300 173.1933 813
#> 13     158  14 33.95   1900 173.1933 813
#> 14     158 127 11.35    750 173.1933 813
#> 15     158 127  8.75    750 173.1933 813
#> 16     158 127 10.50    750 173.1933 813
#> 
#> $shrubData
#>   Species Cover Height       Z50   Z95
#> 1     174     2     30 1193.7766 10000
#> 2     165     3    100  305.4645  1700
#> 3     150     2    120  183.9106   879
#> 4     195     3    100  203.0918  1000
#> 5     216     2    150  183.9106   879
#> 6     109     3     50  144.4250   642
#> 7     133     3     30  105.7274   428
#> 8     191    10    150  151.6395   684
#> 9      75     2     30  153.5119   695
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"  

# Builds a list whose elements are 'forest' objects
l = IFN2forestlist(exampletreedata, exampleshrubdata, SpParamsMED)

#> Number of plots: 5
#> Filtered records in tree data: 20
#> Translating species codes...
#> Shrub data records with unrecognized IFN species codes: 3/43 (7%)
#> 
#> 141 
#>   3 
#> Extracting IFN data...
#> done.

names(l)

#> [1] "81065" "81066" "81067" "81068" "81069"

l[[1]]

#> $ID
#> [1] "81065"
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#>    Species   N   DBH Height      Z50 Z95
#> 1      158 127  9.55    750 173.1933 813
#> 2      158 127  9.40    650 173.1933 813
#> 3      158  32 21.00   1600 173.1933 813
#> 4      158  14 23.30   1450 173.1933 813
#> 5      158  32 22.15   1450 173.1933 813
#> 6      158  14 23.10   1450 173.1933 813
#> 7      158  32 19.00   1350 173.1933 813
#> 8      158  14 25.40   1450 173.1933 813
#> 9      158  32 18.50   1250 173.1933 813
#> 10     158 127  8.20    750 173.1933 813
#> 11     158  32 21.25   1400 173.1933 813
#> 12     158  32 19.30   1300 173.1933 813
#> 13     158  14 33.95   1900 173.1933 813
#> 14     158 127 11.35    750 173.1933 813
#> 15     158 127  8.75    750 173.1933 813
#> 16     158 127 10.50    750 173.1933 813
#> 
#> $shrubData
#>   Species Cover Height       Z50   Z95
#> 1     174     2     30 1193.7766 10000
#> 2     165     3    100  305.4645  1700
#> 3     150     2    120  183.9106   879
#> 4     195     3    100  203.0918  1000
#> 5     216     2    150  183.9106   879
#> 6     109     3     50  144.4250   642
#> 7     133     3     30  105.7274   428
#> 8     191    10    150  151.6395   684
#> 9      75     2     30  153.5119   695
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"