Estimate Rooting Depth Parameters

Estimates rooting depth parameters for forest objects following Cabon et al. (2018).

estimateRootingDepth(
  forest,
  soil,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

estimateRootingDepthList(
  forestlist,
  soillist,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

Arguments

forest

An object of class forest.

soil

An object of class soil or a data frame with soil parameters.

PET_summer

Average (climatic) potential evapotranspiration (mm) corresponding to summer months.

P_summer

Average (climatic) precipitation (mm) corresponding to summer months.

SpParams

A data frame with species parameters (see SpParamsMED).

fillMissing

A boolean flag to indicate that missing estimates should be filled with mean values.

forestlist

A list of objects of class forest.

soillist

A list with objects of class soil or data frames with soil parameters.

Value

Function estimateRootingDepth returns object of class forest with modified values for Z50 and Z95. Function estimateRootingDepthList returns a list of forest objects with modified root distribution parameters.

Details

If PET_summer or P_summer are of length one, their elements are repeated for all elements in forestlist. Similarly, if soillist is an object of class soil or a data frame with soil parameters, the same soil parameters are repeated for all elements in forestlist.

References

Cabon, A., Martinez-Vilalta, J., de Aragon, J.M., Poyatos, R., De Caceres, M., 2018. Applying the eco-hydrological equilibrium hypothesis to model root distribution in water-limited forests. Ecohydrology 11, e2015. https://doi.org/10.1002/eco.2015

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

# Builds from IFN3 data a list whose elements are 'forest' objects
data(piesMayoresIFN3)
data(regeneraIFN3)
data(matorralIFN3)
l1 <- IFN2forest(piesMayoresIFN3, SpParamsMED,
               matorral = matorralIFN3, regenera = regeneraIFN3)
#> Number of plots: 5
#> Filtered records in tree data: 20
#> Translating species codes...
#> Extracting IFN data...
#> done.

# Estimate rooting depth
soilpar <- defaultSoilParams()
l2 <- estimateRootingDepthList(l1, soilpar, 300, 100, SpParamsMED)

# Compare tree rooting depths
print(l1[[1]]$treeData)
#>    Species           N  DBH Height      Z50  Z95 OrdenIf2 OrdenIf3
#> 1      149   31.830989 19.1    750 173.1933  813      000      001
#> 2      149   31.830989 18.8    650 173.1933  813      000      002
#> 3      149   14.147106 42.0   1600 173.1933  813      002      003
#> 4      149    5.092958 46.6   1450 173.1933  813      000      004
#> 5      149    5.092958 44.3   1450 173.1933  813      004      005
#> 6      149    5.092958 46.2   1450 173.1933  813      000      006
#> 7      149   14.147106 38.0   1350 173.1933  813      005      007
#> 8      149    5.092958 50.8   1450 173.1933  813      009      008
#> 9      149   14.147106 37.0   1250 173.1933  813      011      009
#> 10     149   31.830989 16.4    750 173.1933  813      000      010
#> 11     149    5.092958 42.5   1400 173.1933  813      013      011
#> 12     149   14.147106 38.6   1300 173.1933  813      014      012
#> 13     149    5.092958 67.9   1900 173.1933  813      017      013
#> 14     149   14.147106 22.7    750 173.1933  813      000      014
#> 15     149   31.830989 17.5    750 173.1933  813      000      015
#> 16     149   31.830989 21.0    750 173.1933  813      000      016
#> 17     149 1400.563501  5.0    350 173.1933  813     <NA>     <NA>
#> 18     149  318.309887  0.1     10 173.1933  813     <NA>     <NA>
#> 19     149 1273.239546  0.5     80 173.1933  813     <NA>     <NA>
#> 20     149 1273.239546  1.5    100 173.1933  813     <NA>     <NA>
#> 21     115  318.309887  0.5     80 199.9600  980     <NA>     <NA>
#> 22     166 1273.239546  0.1     10 647.0011 4510     <NA>     <NA>
#> 23     166  318.309887  0.5     80 647.0011 4510     <NA>     <NA>
#> 24     168  318.309887  0.1     10 702.5790 5020     <NA>     <NA>
print(l2[[1]]$treeData)
#>    Species           N  DBH Height      Z50     Z95 OrdenIf2 OrdenIf3
#> 1      149   31.830989 19.1    750 203.6321 500.000      000      001
#> 2      149   31.830989 18.8    650 203.6321 500.000      000      002
#> 3      149   14.147106 42.0   1600 203.6321 500.000      002      003
#> 4      149    5.092958 46.6   1450 203.6321 500.000      000      004
#> 5      149    5.092958 44.3   1450 203.6321 500.000      004      005
#> 6      149    5.092958 46.2   1450 203.6321 500.000      000      006
#> 7      149   14.147106 38.0   1350 203.6321 500.000      005      007
#> 8      149    5.092958 50.8   1450 203.6321 500.000      009      008
#> 9      149   14.147106 37.0   1250 203.6321 500.000      011      009
#> 10     149   31.830989 16.4    750 203.6321 500.000      000      010
#> 11     149    5.092958 42.5   1400 203.6321 500.000      013      011
#> 12     149   14.147106 38.6   1300 203.6321 500.000      014      012
#> 13     149    5.092958 67.9   1900 203.6321 500.000      017      013
#> 14     149   14.147106 22.7    750 203.6321 500.000      000      014
#> 15     149   31.830989 17.5    750 203.6321 500.000      000      015
#> 16     149   31.830989 21.0    750 203.6321 500.000      000      016
#> 17     149 1400.563501  5.0    350 203.6321 500.000     <NA>     <NA>
#> 18     149  318.309887  0.1     10 203.6321 500.000     <NA>     <NA>
#> 19     149 1273.239546  0.5     80 203.6321 500.000     <NA>     <NA>
#> 20     149 1273.239546  1.5    100 203.6321 500.000     <NA>     <NA>
#> 21     115  318.309887  0.5     80 286.2504 500.000     <NA>     <NA>
#> 22     166 1273.239546  0.1     10 249.7409 562.058     <NA>     <NA>
#> 23     166  318.309887  0.5     80 249.7409 562.058     <NA>     <NA>
#> 24     168  318.309887  0.1     10 293.7324 500.000     <NA>     <NA>