Example forest inventory data

Example data to illustrate the creation of forest objects from inventory data

Source

Data tables piesMayoresIFN2, piesMenoresIFN2 and matorralIFN2 were extracted from the Second Spanish National Forest Inventory.

Data tables piesMayoresIFN3, regeneraIFN3 and matorralIFN3 were extracted from the Third Spanish National Forest Inventory.

Data table poblet_trees corresponds to field data sampled by the Catalan Forest Ownership Center (Centre de la Propietat Forestal; CPF).

Details

Six example data tables were extracted from the second and third surveys of the Spanish National Forest Inventory (IFN2 and IFN3), to illustrate function IFN2forest:

piesMayoresIFN2 - Data frame with tree data (79 observations and 9 variables) from IFN2.

Provincia - Spanish province code (character)

Estadillo - Plot code within province (character)

ID - Plot ID, resulting from the concatenation of 'Provincia' and 'Estadillo' (character)

OrdenIf2 - Tree code in the plot (character)

Especie - Species code in IFN2 (character)

Dn1: Diameter at breast height (in mm), first measurement.

Dn2: Diameter at breast height (in mm), second measurement.

Ht: Total tree height (in m)

piesMenoresIFN2 - Data frame with tree regeneration data (27 observations and 7 variables) from IFN2.

Provincia - Spanish province code (character)

Estadillo - Plot code within province (character)

ID - Plot ID, resulting from the concatenation of 'Provincia' and 'Estadillo' (character)

Especie - Species code in IFN2 (character)

Numero - Number of stems for 2.5-7.5 cm DBH class (integer)

Hm - Average tree height (dm)

Regena - Regeneration code ('0', 1', '2', '3') (integer)

matorralIFN2 - Data frame with shrub data (20 observations and 6 variables) from IFN2.

Provincia - Spanish province code (character)

Estadillo - Plot code within province (character)

ID - Plot ID, resulting from the concatenation of 'Provincia' and 'Estadillo' (character)

Especie - Species code in IFN2 (character)

Fcc - Shrub cover value (%)

Hm - Average shrub height (dm)

piesMayoresIFN3 - Data frame with tree data (118 observations and 13 variables) from IFN3.

Provincia - Spanish province code (character)

Estadillo - Plot code within province (character)

Clase - Plot type (character)

Subclase - Plot subtype (character)

IDPARCELA - Plot ID, resulting from the concatenation of 'Provincia' and 'Estadillo' (character)

IDCLASE - Plot type ID, resulting from the concatenation of 'Clase' and 'SubClase' (character)

ID - Unique plot ID, resulting from the concatenation of 'IDPARCELA' and 'IDCLASE' (character)

OrdenIf2 - Tree code in the plot during IFN2 (character)

OrdenIf3 - Tree code in the plot during IFN3 (character)

Especie - Species code in IFN3 (character)

Dn1: Diameter at breast height (in mm), first measurement.

Dn2: Diameter at breast height (in mm), second measurement.

Ht: Total tree height (in m)

regeneraIFN3 - Data frame with tree regeneration data (86 observations and 13 variables) from IFN3.

Provincia - Spanish province code (character)

Estadillo - Plot code within province (character)

Clase - Plot type (character)

Subclase - Plot subtype (character)

IDPARCELA - Plot ID, resulting from the concatenation of 'Provincia' and 'Estadillo' (character)

IDCLASE - Plot type ID, resulting from the concatenation of 'Clase' and 'SubClase' (character)

ID - Unique plot ID, resulting from the concatenation of 'IDPARCELA' and 'IDCLASE' (character)

Especie - Species code in IFN3 (character)

CatDes - Regeneration category, either '1', '2', '3' or '4' (character)

Tipo - Kind of regeneration (character)

Densidad - Density category ( 1', '2', or '3') (character)

NumPies - Number of stems (for CatDes == '1') (numeric)

matorralIFN3 - Data frame with shrub data (43 observations and 10 variables) from IFN3.

Provincia - Spanish province code (character)

Estadillo - Plot code within province (character)

Clase - Plot type (character)

Subclase - Plot subtype (character)

IDPARCELA - Plot ID, resulting from the concatenation of 'Provincia' and 'Estadillo' (character)

IDCLASE - Plot type ID, resulting from the concatenation of 'Clase' and 'SubClase' (character)

ID - Unique plot ID, resulting from the concatenation of 'IDPARCELA' and 'IDCLASE' (character)

Especie - Species code in IFN3 (character)

Fcc - Shrub cover value (%)

Hm - Average shrub height (dm)

An additional data table comes from an independent forest inventory survey, used to illustrate the general function forest_mapTreeTable:

poblet_trees - Data frame with example tree plot data from Poblet, Catalonia (717 observations and 4 variables).

Plot.Code - Plot ID (character)

Indv.Ref - Tree individual (integer)

Species - Species name (character)

Diameter.cm - Tree diameter at breast height (cm)