Map forest plot data

Mapping functions to facilitate building forest objects from forest plot data

forest_mapTreeTable(x, mapping, SpParams, plot.size = NULL)

forest_mapShrubTable(y, mapping, SpParams, plot.size = NULL)

forest_mapWoodyTables(x, y, mapping, SpParams, plot.size = NULL)

Arguments

x

A data frame with tree records in rows and attributes in columns

mapping

A named character vector to specify mappings of columns in x or y into attributes of treeData and shrubData data frames. Accepted names (and the corresponding specifications for the columns in x and y) are:

"Species": Species code (should follow codes in SpParams).

"Species.name": Species name. In this case, the species code will be drawn by matching names with species names in SpParams.

"N": Tree density (in ind./ha).

"Cover": Shrub cover (in %).

"D1": Shrub largest crown diameter (in cm).

"D2": Shrub crown diameter orthogonal to the largest one (in cm).

"plot.size": Plot size (in m2) to which each plot record refers to.

"DBH": Diameter at breast height (in cm).

"Height": Tree or shrub height (in cm).

"Z50": Depth (in mm) corresponding to 50% of fine roots.

"Z95": Depth (in mm) corresponding to 95% of fine roots.

SpParams

A data frame with species parameters (see SpParamsMED) from which valid species names are drawn.

plot.size

The size of plot sampled area (in m2). Alternatively, 'plot.size' can be a column in x or y and specified in mapping to indicate that trees/shrubs have been measured in different subplots and, therefore, they represent different densities per hectare or cover values.

y

A data frame with shrub records in rows and attributes in columns

Value

Functions forest_mapTreeTable and forest_mapShrubTable return a data frame with the structure of treeData and shrubData from forest objects. Function forest_mapWoodyTable returns directly a forest object.

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

# Load species parameters
data(SpParamsMED)

# (1) Mapping tree data
# Load Poblet tree data
data(poblet_trees)

# Subset control plot
x <- subset(poblet_trees, Plot.Code=="POBL_CTL")

# Estimate sampled area (15-m radius plot)
sampled_area <- pi*15^2

# Define mapping
mapping_x <- c("Species.name" = "Species", "DBH" = "Diameter.cm")

# Map tree data for plot 'POBL_CTL'
forest_mapTreeTable(x,
                    mapping = mapping_x, SpParams = SpParamsMED,
                    plot.size = sampled_area)
#>     Species        N Height  DBH Z50 Z95
#> 1         4 14.14711     NA  7.6  NA  NA
#> 2        19 14.14711     NA  7.5  NA  NA
#> 3        19 14.14711     NA  7.5  NA  NA
#> 4        19 14.14711     NA  7.5  NA  NA
#> 5        19 14.14711     NA  7.5  NA  NA
#> 6        19 14.14711     NA  7.5  NA  NA
#> 7        19 14.14711     NA  7.5  NA  NA
#> 8        19 14.14711     NA  7.7  NA  NA
#> 9        19 14.14711     NA  7.8  NA  NA
#> 10       19 14.14711     NA  7.8  NA  NA
#> 11       19 14.14711     NA  7.8  NA  NA
#> 12       19 14.14711     NA  7.8  NA  NA
#> 13       19 14.14711     NA  7.8  NA  NA
#> 14       19 14.14711     NA  7.9  NA  NA
#> 15       19 14.14711     NA  8.0  NA  NA
#> 16       19 14.14711     NA  8.0  NA  NA
#> 17       19 14.14711     NA  8.0  NA  NA
#> 18       19 14.14711     NA  8.0  NA  NA
#> 19       19 14.14711     NA  8.1  NA  NA
#> 20       19 14.14711     NA  8.2  NA  NA
#> 21       19 14.14711     NA  8.2  NA  NA
#> 22       19 14.14711     NA  8.2  NA  NA
#> 23       19 14.14711     NA  8.3  NA  NA
#> 24       19 14.14711     NA  8.3  NA  NA
#> 25       19 14.14711     NA  8.4  NA  NA
#> 26       19 14.14711     NA  8.4  NA  NA
#> 27       19 14.14711     NA  8.5  NA  NA
#> 28       19 14.14711     NA  8.5  NA  NA
#> 29       19 14.14711     NA  8.5  NA  NA
#> 30       19 14.14711     NA  8.5  NA  NA
#> 31       19 14.14711     NA  8.5  NA  NA
#> 32       19 14.14711     NA  8.6  NA  NA
#> 33       19 14.14711     NA  8.6  NA  NA
#> 34       19 14.14711     NA  8.7  NA  NA
#> 35       19 14.14711     NA  8.7  NA  NA
#> 36       19 14.14711     NA  8.8  NA  NA
#> 37       19 14.14711     NA  8.8  NA  NA
#> 38       19 14.14711     NA  8.9  NA  NA
#> 39       19 14.14711     NA  9.0  NA  NA
#> 40       19 14.14711     NA  9.0  NA  NA
#> 41       19 14.14711     NA  9.0  NA  NA
#> 42       19 14.14711     NA  9.0  NA  NA
#> 43       19 14.14711     NA  9.0  NA  NA
#> 44       19 14.14711     NA  9.1  NA  NA
#> 45       19 14.14711     NA  9.2  NA  NA
#> 46       19 14.14711     NA  9.3  NA  NA
#> 47       19 14.14711     NA  9.4  NA  NA
#> 48       19 14.14711     NA  9.4  NA  NA
#> 49       19 14.14711     NA  9.4  NA  NA
#> 50       19 14.14711     NA  9.4  NA  NA
#> 51       19 14.14711     NA  9.5  NA  NA
#> 52       19 14.14711     NA  9.5  NA  NA
#> 53       19 14.14711     NA  9.5  NA  NA
#> 54       19 14.14711     NA  9.5  NA  NA
#> 55       19 14.14711     NA  9.5  NA  NA
#> 56       19 14.14711     NA  9.6  NA  NA
#> 57       19 14.14711     NA  9.7  NA  NA
#> 58       19 14.14711     NA  9.7  NA  NA
#> 59       19 14.14711     NA  9.8  NA  NA
#> 60       19 14.14711     NA  9.8  NA  NA
#> 61       19 14.14711     NA  9.8  NA  NA
#> 62       19 14.14711     NA  9.8  NA  NA
#> 63       19 14.14711     NA 10.0  NA  NA
#> 64       19 14.14711     NA 10.0  NA  NA
#> 65       19 14.14711     NA 10.0  NA  NA
#> 66       19 14.14711     NA 10.1  NA  NA
#> 67       19 14.14711     NA 10.1  NA  NA
#> 68       19 14.14711     NA 10.1  NA  NA
#> 69       19 14.14711     NA 10.2  NA  NA
#> 70       19 14.14711     NA 10.2  NA  NA
#> 71       19 14.14711     NA 10.4  NA  NA
#> 72       19 14.14711     NA 10.5  NA  NA
#> 73       19 14.14711     NA 10.8  NA  NA
#> 74       19 14.14711     NA 10.8  NA  NA
#> 75       19 14.14711     NA 10.8  NA  NA
#> 76       19 14.14711     NA 10.9  NA  NA
#> 77       19 14.14711     NA 11.0  NA  NA
#> 78       19 14.14711     NA 11.0  NA  NA
#> 79       19 14.14711     NA 11.0  NA  NA
#> 80       19 14.14711     NA 11.0  NA  NA
#> 81       19 14.14711     NA 11.1  NA  NA
#> 82       19 14.14711     NA 11.2  NA  NA
#> 83       19 14.14711     NA 11.4  NA  NA
#> 84       19 14.14711     NA 11.4  NA  NA
#> 85       19 14.14711     NA 11.5  NA  NA
#> 86       19 14.14711     NA 11.7  NA  NA
#> 87       19 14.14711     NA 11.7  NA  NA
#> 88       19 14.14711     NA 11.8  NA  NA
#> 89       19 14.14711     NA 12.0  NA  NA
#> 90       19 14.14711     NA 12.2  NA  NA
#> 91       19 14.14711     NA 12.2  NA  NA
#> 92       19 14.14711     NA 12.3  NA  NA
#> 93       19 14.14711     NA 12.5  NA  NA
#> 94       19 14.14711     NA 12.5  NA  NA
#> 95       19 14.14711     NA 12.5  NA  NA
#> 96       19 14.14711     NA 12.6  NA  NA
#> 97       19 14.14711     NA 12.8  NA  NA
#> 98       19 14.14711     NA 13.0  NA  NA
#> 99       19 14.14711     NA 13.0  NA  NA
#> 100      19 14.14711     NA 13.3  NA  NA
#> 101      19 14.14711     NA 13.3  NA  NA
#> 102      19 14.14711     NA 13.3  NA  NA
#> 103      19 14.14711     NA 13.6  NA  NA
#> 104      19 14.14711     NA 13.8  NA  NA
#> 105      19 14.14711     NA 13.9  NA  NA
#> 106      19 14.14711     NA 13.9  NA  NA
#> 107      19 14.14711     NA 14.0  NA  NA
#> 108      19 14.14711     NA 14.4  NA  NA
#> 109      19 14.14711     NA 14.4  NA  NA
#> 110      19 14.14711     NA 15.2  NA  NA
#> 111      19 14.14711     NA 17.2  NA  NA
#> 112      19 14.14711     NA 18.9  NA  NA
#> 113     142 14.14711     NA 11.1  NA  NA
#> 114     168 14.14711     NA  7.6  NA  NA
#> 115     168 14.14711     NA  7.6  NA  NA
#> 116     168 14.14711     NA  7.7  NA  NA
#> 117     168 14.14711     NA  7.7  NA  NA
#> 118     168 14.14711     NA  7.7  NA  NA
#> 119     168 14.14711     NA  7.7  NA  NA
#> 120     168 14.14711     NA  7.8  NA  NA
#> 121     168 14.14711     NA  7.8  NA  NA
#> 122     168 14.14711     NA  7.9  NA  NA
#> 123     168 14.14711     NA  8.0  NA  NA
#> 124     168 14.14711     NA  8.1  NA  NA
#> 125     168 14.14711     NA  8.1  NA  NA
#> 126     168 14.14711     NA  8.2  NA  NA
#> 127     168 14.14711     NA  8.3  NA  NA
#> 128     168 14.14711     NA  8.3  NA  NA
#> 129     168 14.14711     NA  8.4  NA  NA
#> 130     168 14.14711     NA  8.4  NA  NA
#> 131     168 14.14711     NA  8.5  NA  NA
#> 132     168 14.14711     NA  8.7  NA  NA
#> 133     168 14.14711     NA  8.8  NA  NA
#> 134     168 14.14711     NA  8.8  NA  NA
#> 135     168 14.14711     NA  8.8  NA  NA
#> 136     168 14.14711     NA  8.9  NA  NA
#> 137     168 14.14711     NA  9.0  NA  NA
#> 138     168 14.14711     NA  9.0  NA  NA
#> 139     168 14.14711     NA  9.0  NA  NA
#> 140     168 14.14711     NA  9.1  NA  NA
#> 141     168 14.14711     NA  9.3  NA  NA
#> 142     168 14.14711     NA  9.3  NA  NA
#> 143     168 14.14711     NA  9.3  NA  NA
#> 144     168 14.14711     NA  9.3  NA  NA
#> 145     168 14.14711     NA  9.3  NA  NA
#> 146     168 14.14711     NA  9.3  NA  NA
#> 147     168 14.14711     NA  9.4  NA  NA
#> 148     168 14.14711     NA  9.4  NA  NA
#> 149     168 14.14711     NA  9.5  NA  NA
#> 150     168 14.14711     NA  9.5  NA  NA
#> 151     168 14.14711     NA  9.5  NA  NA
#> 152     168 14.14711     NA  9.6  NA  NA
#> 153     168 14.14711     NA  9.7  NA  NA
#> 154     168 14.14711     NA  9.7  NA  NA
#> 155     168 14.14711     NA  9.8  NA  NA
#> 156     168 14.14711     NA  9.8  NA  NA
#> 157     168 14.14711     NA 10.0  NA  NA
#> 158     168 14.14711     NA 10.0  NA  NA
#> 159     168 14.14711     NA 10.0  NA  NA
#> 160     168 14.14711     NA 10.0  NA  NA
#> 161     168 14.14711     NA 10.1  NA  NA
#> 162     168 14.14711     NA 10.1  NA  NA
#> 163     168 14.14711     NA 10.2  NA  NA
#> 164     168 14.14711     NA 10.2  NA  NA
#> 165     168 14.14711     NA 10.2  NA  NA
#> 166     168 14.14711     NA 10.2  NA  NA
#> 167     168 14.14711     NA 10.3  NA  NA
#> 168     168 14.14711     NA 10.3  NA  NA
#> 169     168 14.14711     NA 10.3  NA  NA
#> 170     168 14.14711     NA 10.4  NA  NA
#> 171     168 14.14711     NA 10.4  NA  NA
#> 172     168 14.14711     NA 10.5  NA  NA
#> 173     168 14.14711     NA 10.5  NA  NA
#> 174     168 14.14711     NA 10.7  NA  NA
#> 175     168 14.14711     NA 10.7  NA  NA
#> 176     168 14.14711     NA 10.7  NA  NA
#> 177     168 14.14711     NA 10.8  NA  NA
#> 178     168 14.14711     NA 10.9  NA  NA
#> 179     168 14.14711     NA 11.0  NA  NA
#> 180     168 14.14711     NA 11.0  NA  NA
#> 181     168 14.14711     NA 11.2  NA  NA
#> 182     168 14.14711     NA 11.2  NA  NA
#> 183     168 14.14711     NA 11.3  NA  NA
#> 184     168 14.14711     NA 11.4  NA  NA
#> 185     168 14.14711     NA 11.5  NA  NA
#> 186     168 14.14711     NA 11.5  NA  NA
#> 187     168 14.14711     NA 11.5  NA  NA
#> 188     168 14.14711     NA 11.7  NA  NA
#> 189     168 14.14711     NA 11.7  NA  NA
#> 190     168 14.14711     NA 11.7  NA  NA
#> 191     168 14.14711     NA 12.0  NA  NA
#> 192     168 14.14711     NA 12.0  NA  NA
#> 193     168 14.14711     NA 12.0  NA  NA
#> 194     168 14.14711     NA 12.0  NA  NA
#> 195     168 14.14711     NA 12.0  NA  NA
#> 196     168 14.14711     NA 12.2  NA  NA
#> 197     168 14.14711     NA 12.2  NA  NA
#> 198     168 14.14711     NA 12.3  NA  NA
#> 199     168 14.14711     NA 12.5  NA  NA
#> 200     168 14.14711     NA 12.5  NA  NA
#> 201     168 14.14711     NA 12.6  NA  NA
#> 202     168 14.14711     NA 12.6  NA  NA
#> 203     168 14.14711     NA 12.7  NA  NA
#> 204     168 14.14711     NA 12.7  NA  NA
#> 205     168 14.14711     NA 12.8  NA  NA
#> 206     168 14.14711     NA 12.9  NA  NA
#> 207     168 14.14711     NA 13.0  NA  NA
#> 208     168 14.14711     NA 13.1  NA  NA
#> 209     168 14.14711     NA 13.1  NA  NA
#> 210     168 14.14711     NA 13.2  NA  NA
#> 211     168 14.14711     NA 13.2  NA  NA
#> 212     168 14.14711     NA 13.3  NA  NA
#> 213     168 14.14711     NA 13.4  NA  NA
#> 214     168 14.14711     NA 13.5  NA  NA
#> 215     168 14.14711     NA 13.5  NA  NA
#> 216     168 14.14711     NA 13.5  NA  NA
#> 217     168 14.14711     NA 13.5  NA  NA
#> 218     168 14.14711     NA 13.8  NA  NA
#> 219     168 14.14711     NA 13.8  NA  NA
#> 220     168 14.14711     NA 14.0  NA  NA
#> 221     168 14.14711     NA 14.0  NA  NA
#> 222     168 14.14711     NA 14.1  NA  NA
#> 223     168 14.14711     NA 14.1  NA  NA
#> 224     168 14.14711     NA 14.2  NA  NA
#> 225     168 14.14711     NA 14.4  NA  NA
#> 226     168 14.14711     NA 14.6  NA  NA
#> 227     168 14.14711     NA 14.8  NA  NA
#> 228     168 14.14711     NA 14.8  NA  NA
#> 229     168 14.14711     NA 14.8  NA  NA
#> 230     168 14.14711     NA 14.9  NA  NA
#> 231     168 14.14711     NA 15.0  NA  NA
#> 232     168 14.14711     NA 15.0  NA  NA
#> 233     168 14.14711     NA 15.0  NA  NA
#> 234     168 14.14711     NA 15.1  NA  NA
#> 235     168 14.14711     NA 15.2  NA  NA
#> 236     168 14.14711     NA 15.4  NA  NA
#> 237     168 14.14711     NA 15.5  NA  NA
#> 238     168 14.14711     NA 15.5  NA  NA
#> 239     168 14.14711     NA 15.7  NA  NA
#> 240     168 14.14711     NA 15.8  NA  NA
#> 241     168 14.14711     NA 16.2  NA  NA
#> 242     168 14.14711     NA 16.4  NA  NA
#> 243     168 14.14711     NA 16.4  NA  NA
#> 244     168 14.14711     NA 16.5  NA  NA
#> 245     168 14.14711     NA 16.5  NA  NA
#> 246     168 14.14711     NA 16.5  NA  NA
#> 247     168 14.14711     NA 16.5  NA  NA
#> 248     168 14.14711     NA 16.5  NA  NA
#> 249     168 14.14711     NA 16.9  NA  NA
#> 250     168 14.14711     NA 17.0  NA  NA
#> 251     168 14.14711     NA 17.2  NA  NA
#> 252     168 14.14711     NA 17.2  NA  NA
#> 253     168 14.14711     NA 17.3  NA  NA
#> 254     168 14.14711     NA 17.3  NA  NA
#> 255     168 14.14711     NA 17.5  NA  NA
#> 256     168 14.14711     NA 17.5  NA  NA
#> 257     168 14.14711     NA 18.3  NA  NA
#> 258     168 14.14711     NA 18.3  NA  NA
#> 259     168 14.14711     NA 18.7  NA  NA
#> 260     168 14.14711     NA 18.9  NA  NA
#> 261     168 14.14711     NA 20.0  NA  NA
#> 262     168 14.14711     NA 20.4  NA  NA
#> 263     168 14.14711     NA 20.7  NA  NA
#> 264     168 14.14711     NA 22.2  NA  NA
#> 265     168 14.14711     NA 22.4  NA  NA
#> 266     168 14.14711     NA 25.4  NA  NA
#> 267     168 14.14711     NA 26.0  NA  NA

# (2) Mapping shrub individual data
#
# Create the individual shrub data frame
species <- c("Erica arborea","Cistus albidus", "Erica arborea", "Cistus albidus", "Cistus albidus")
H <- c(200,50,100,40,30)
D1 <- c(140,40,100, 35,30)
D2 <- D1
y <- data.frame(species, H, D1, D2)

# Define mapping (D1 and D2 map to variables with the same name)
mapping_y <- c("Species.name"= "species", "Height" ="H", "D1", "D2")

# Map individual shrub data to cover data (here each individual becomes a cohort)
# assuming that the sampled area was 4 m2
forest_mapShrubTable(y,
                     mapping = mapping_y, SpParams = SpParamsMED,
                     plot.size = 4)
#>   Species Height     Cover Z50 Z95
#> 1      83    200 38.484510  NA  NA
#> 2      40     50  3.141593  NA  NA
#> 3      83    100 19.634954  NA  NA
#> 4      40     40  2.405282  NA  NA
#> 5      40     30  1.767146  NA  NA