Extract forest from IFN data

Creates a forest object from Spanish Forest Inventory (IFN) data (DGCN 2005).

IFN2forest(
  IFNtreeData,
  IFNshrubData,
  ID,
  SpParams,
  IFNherbData = NULL,
  patchsize = 10000,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  keepNumOrden = TRUE,
  verbose = TRUE
)

IFN2forestlist(
  IFNtreeData,
  IFNshrubData,
  SpParams,
  IFNherbData = NULL,
  setDefaults = TRUE,
  filterWrongRecords = TRUE,
  keepNumOrden = TRUE,
  verbose = TRUE
)

Arguments

IFNtreeData

A data frame with measured tree data.

IFNshrubData

A data frame with measured shrub data.

ID

A string with the ID of the plot to be extracted.

SpParams

A data frame with species parameters (see SpParamsMED).

IFNherbData

A data frame with cover and mean height of the herb layer.

patchsize

The area of the forest stand, in square meters.

setDefaults

Initializes default values for missing fields in IFN data.

filterWrongRecords

Filters wrong records (records with missing values, zero values or wrong growth forms)

keepNumOrden

Keeps num orden as additional column (OIF2, OIF3, ...)

verbose

A boolean flag to indicate console output.

Value

Function IFN2forest returns an object of class forest, whereas function IFN2forestlist returns a list of forest objects.

Details

IFN input data needs to be in a specific format, see package IFNread. Functions IFN2forest and IFN2forestlist call translateIFNSpeciesCodes internally.

References

DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

data(SpParamsMED)
data(exampletreedata)
data(exampleshrubdata)

# Builds an object 'forest' corresponding to one specific forest plot
f = IFN2forest(exampletreedata, exampleshrubdata,
               ID = "81065", SpParams = SpParamsMED)
#> Filtered records in tree data: 11
print(f)
#> $ID
#> [1] "81065"
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#>    Species         N   DBH Height      Z50 Z95 OIF2 OIF3
#> 1      149 127.32395  9.55    750 173.1933 813    0    1
#> 2      149 127.32395  9.40    650 173.1933 813    0    2
#> 3      149  31.83099 21.00   1600 173.1933 813    2    3
#> 4      149  14.14711 23.30   1450 173.1933 813    0    4
#> 5      149  31.83099 22.15   1450 173.1933 813    4    5
#> 6      149  14.14711 23.10   1450 173.1933 813    0    6
#> 7      149  31.83099 19.00   1350 173.1933 813    5    7
#> 8      149  14.14711 25.40   1450 173.1933 813    9    8
#> 9      149  31.83099 18.50   1250 173.1933 813   11    9
#> 10     149 127.32395  8.20    750 173.1933 813    0   10
#> 11     149  31.83099 21.25   1400 173.1933 813   13   11
#> 12     149  31.83099 19.30   1300 173.1933 813   14   12
#> 13     149  14.14711 33.95   1900 173.1933 813   17   13
#> 14     149 127.32395 11.35    750 173.1933 813    0   14
#> 15     149 127.32395  8.75    750 173.1933 813    0   15
#> 16     149 127.32395 10.50    750 173.1933 813    0   16
#> 
#> $shrubData
#>    Species Cover Height       Z50   Z95
#> 1      165     2     30 1193.7766 10000
#> 2      156     3    100  305.4645  1700
#> 3      141     2    120  191.2703   925
#> 4      190     3    100  203.0918  1000
#> 5      214     2    150  191.2703   925
#> 6      107     2     10  191.2703   925
#> 7      101     3     50  143.9056   639
#> 8      123     3     30  105.7274   428
#> 9      183    10    150  151.6395   684
#> 10      79     2     30  153.5119   695
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"  

# Builds a list whose elements are 'forest' objects
l = IFN2forestlist(exampletreedata, exampleshrubdata, SpParamsMED)
#> Number of plots: 5
#> Filtered records in tree data: 20
#> Translating species codes...
#> Extracting IFN data...
#> done.
names(l)
#> [1] "81065" "81066" "81067" "81068" "81069"
l[[1]]
#> $ID
#> [1] "81065"
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#>    Species         N   DBH Height      Z50 Z95 OIF2 OIF3
#> 1      149 127.32395  9.55    750 173.1933 813    0    1
#> 2      149 127.32395  9.40    650 173.1933 813    0    2
#> 3      149  31.83099 21.00   1600 173.1933 813    2    3
#> 4      149  14.14711 23.30   1450 173.1933 813    0    4
#> 5      149  31.83099 22.15   1450 173.1933 813    4    5
#> 6      149  14.14711 23.10   1450 173.1933 813    0    6
#> 7      149  31.83099 19.00   1350 173.1933 813    5    7
#> 8      149  14.14711 25.40   1450 173.1933 813    9    8
#> 9      149  31.83099 18.50   1250 173.1933 813   11    9
#> 10     149 127.32395  8.20    750 173.1933 813    0   10
#> 11     149  31.83099 21.25   1400 173.1933 813   13   11
#> 12     149  31.83099 19.30   1300 173.1933 813   14   12
#> 13     149  14.14711 33.95   1900 173.1933 813   17   13
#> 14     149 127.32395 11.35    750 173.1933 813    0   14
#> 15     149 127.32395  8.75    750 173.1933 813    0   15
#> 16     149 127.32395 10.50    750 173.1933 813    0   16
#> 
#> $shrubData
#>    Species Cover Height       Z50   Z95
#> 1      165     2     30 1193.7766 10000
#> 2      156     3    100  305.4645  1700
#> 3      141     2    120  191.2703   925
#> 4      190     3    100  203.0918  1000
#> 5      214     2    150  191.2703   925
#> 6      107     2     10  191.2703   925
#> 7      101     3     50  143.9056   639
#> 8      123     3     30  105.7274   428
#> 9      183    10    150  151.6395   684
#> 10      79     2     30  153.5119   695
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"