Estimate Rooting Depth Parameters

Estimates rooting depth parameters for forest objects following Cabon et al. (2018).

estimateRootingDepth(
  forest,
  soil,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

estimateRootingDepthList(
  forestlist,
  soillist,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

Arguments

forest

An object of class forest.

soil

An object of class soil or a data frame with soil parameters.

PET_summer

Average (climatic) potential evapotranspiration (mm) corresponding to summer months.

P_summer

Average (climatic) precipitation (mm) corresponding to summer months.

SpParams

A data frame with species parameters (see SpParamsMED).

fillMissing

A boolean flag to indicate that missing estimates should be filled with mean values.

forestlist

A list of objects of class forest.

soillist

A list with objects of class soil or data frames with soil parameters.

Value

Function estimateRootingDepth returns object of class forest with modified values for Z50 and Z95. Function estimateRootingDepthList returns a list of forest objects with modified root distribution parameters.

Details

If PET_summer or P_summer are of length one, their elements are repeated for all elements in forestlist. Similarly, if soillist is an object of class soil or a data frame with soil parameters, the same soil parameters are repeated for all elements in forestlist.

References

Cabon, A., Martinez-Vilalta, J., de Aragon, J.M., Poyatos, R., De Caceres, M., 2018. Applying the eco-hydrological equilibrium hypothesis to model root distribution in water-limited forests. Ecohydrology 11, e2015. https://doi.org/10.1002/eco.2015

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

data(SpParamsMED)
data(example_treedata_ifn)
data(example_shrubdata_ifn)

# Builds an object 'forest' corresponding to one specific forest plot
f1 = IFN2forest(example_treedata_ifn, example_shrubdata_ifn,
                ID = "81065", SpParams = SpParamsMED)
#> Filtered records in tree data: 11
print(f1)
#> $treeData
#>    Species         N   DBH Height      Z50 Z95 OIF2 OIF3
#> 1      149 127.32395  9.55    750 173.1933 813    0    1
#> 2      149 127.32395  9.40    650 173.1933 813    0    2
#> 3      149  31.83099 21.00   1600 173.1933 813    2    3
#> 4      149  14.14711 23.30   1450 173.1933 813    0    4
#> 5      149  31.83099 22.15   1450 173.1933 813    4    5
#> 6      149  14.14711 23.10   1450 173.1933 813    0    6
#> 7      149  31.83099 19.00   1350 173.1933 813    5    7
#> 8      149  14.14711 25.40   1450 173.1933 813    9    8
#> 9      149  31.83099 18.50   1250 173.1933 813   11    9
#> 10     149 127.32395  8.20    750 173.1933 813    0   10
#> 11     149  31.83099 21.25   1400 173.1933 813   13   11
#> 12     149  31.83099 19.30   1300 173.1933 813   14   12
#> 13     149  14.14711 33.95   1900 173.1933 813   17   13
#> 14     149 127.32395 11.35    750 173.1933 813    0   14
#> 15     149 127.32395  8.75    750 173.1933 813    0   15
#> 16     149 127.32395 10.50    750 173.1933 813    0   16
#> 
#> $shrubData
#>    Species Cover Height      Z50  Z95
#> 1      165     2     30 647.0011 4510
#> 2      156     3    100 305.4645 1700
#> 3      141     2    120 173.0294  812
#> 4      188     3    100 203.0918 1000
#> 5      212     2    150 173.0294  812
#> 6      107     2     10 173.0294  812
#> 7      101     3     50 143.9056  639
#> 8      123     3     30 105.7274  428
#> 9      183    10    150 151.6395  684
#> 10      79     2     30 153.5119  695
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"  

soilpar = defaultSoilParams()
f2 = estimateRootingDepth(f1, soilpar, 300, 100, SpParamsMED)
print(f2)
#> $treeData
#>    Species         N   DBH Height      Z50 Z95 OIF2 OIF3
#> 1      149 127.32395  9.55    750 208.9364 500    0    1
#> 2      149 127.32395  9.40    650 208.9364 500    0    2
#> 3      149  31.83099 21.00   1600 208.9364 500    2    3
#> 4      149  14.14711 23.30   1450 208.9364 500    0    4
#> 5      149  31.83099 22.15   1450 208.9364 500    4    5
#> 6      149  14.14711 23.10   1450 208.9364 500    0    6
#> 7      149  31.83099 19.00   1350 208.9364 500    5    7
#> 8      149  14.14711 25.40   1450 208.9364 500    9    8
#> 9      149  31.83099 18.50   1250 208.9364 500   11    9
#> 10     149 127.32395  8.20    750 208.9364 500    0   10
#> 11     149  31.83099 21.25   1400 208.9364 500   13   11
#> 12     149  31.83099 19.30   1300 208.9364 500   14   12
#> 13     149  14.14711 33.95   1900 208.9364 500   17   13
#> 14     149 127.32395 11.35    750 208.9364 500    0   14
#> 15     149 127.32395  8.75    750 208.9364 500    0   15
#> 16     149 127.32395 10.50    750 208.9364 500    0   16
#> 
#> $shrubData
#>    Species Cover Height      Z50 Z95
#> 1      165     2     30 250.0000 500
#> 2      156     3    100 250.0000 500
#> 3      141     2    120 250.0000 500
#> 4      188     3    100 250.0000 500
#> 5      212     2    150 236.8821 500
#> 6      107     2     10 172.1502 500
#> 7      101     3     50 203.0155 500
#> 8      123     3     30 250.0000 500
#> 9      183    10    150 172.7408 500
#> 10      79     2     30 203.0155 500
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"  

l1 = IFN2forestlist(example_treedata_ifn, example_shrubdata_ifn, SpParamsMED)
#> Number of plots: 5
#> Filtered records in tree data: 20
#> Translating species codes...
#> Extracting IFN data...
#> done.
l2 = estimateRootingDepthList(l1, soilpar, 300, 100, SpParamsMED)