Estimate Rooting Depth Parameters

Estimates rooting depth parameters for forest objects following Cabon et al. (2018).

estimateRootingDepth(
  forest,
  soil,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

estimateRootingDepthList(
  forestlist,
  soillist,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

Arguments

forest
An object of class forest.

soil
An object of class soil or a data frame with soil parameters.

PET_summer
Average (climatic) potential evapotranspiration (mm) corresponding to summer months.

P_summer
Average (climatic) precipitation (mm) corresponding to summer months.

SpParams
A data frame with species parameters (see SpParamsMED).

fillMissing
A boolean flag to indicate that missing estimates should be filled with mean values.

forestlist
A list of objects of class forest.

soillist
A list with objects of class soil or data frames with soil parameters.

Value

Function estimateRootingDepth returns object of class forest with modified values for Z50 and Z95. Function estimateRootingDepthList returns a list of forest objects with modified root distribution parameters.

Details

If PET_summer or P_summer are of length one, their elements are repeated for all elements in forestlist. Similarly, if soillist is an object of class soil or a data frame with soil parameters, the same soil parameters are repeated for all elements in forestlist.

References

Cabon, A., Martinez-Vilalta, J., de Aragon, J.M., Poyatos, R., De Caceres, M., 2018. Applying the eco-hydrological equilibrium hypothesis to model root distribution in water-limited forests. Ecohydrology 11, e2015. https://doi.org/10.1002/eco.2015

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

data(SpParamsMED)
data(exampletreedata)
data(exampleshrubdata)

# Builds an object 'forest' corresponding to one specific forest plot
f1 = IFN2forest(exampletreedata, exampleshrubdata,
                ID = "81065", SpParams = SpParamsMED)

#> Filtered records in tree data: 11
#> Shrub data records with unrecognized IFN species codes: 1/10
#> 
#> 141 
#>   1 

print(f1)

#> $ID
#> [1] "81065"
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#>    Species   N   DBH Height Z50  Z95
#> 1       55 127  9.55    750 400 1000
#> 2       55 127  9.40    650 400 1000
#> 3       55  32 21.00   1600 400 1000
#> 4       55  14 23.30   1450 400 1000
#> 5       55  32 22.15   1450 400 1000
#> 6       55  14 23.10   1450 400 1000
#> 7       55  32 19.00   1350 400 1000
#> 8       55  14 25.40   1450 400 1000
#> 9       55  32 18.50   1250 400 1000
#> 10      55 127  8.20    750 400 1000
#> 11      55  32 21.25   1400 400 1000
#> 12      55  32 19.30   1300 400 1000
#> 13      55  14 33.95   1900 400 1000
#> 14      55 127 11.35    750 400 1000
#> 15      55 127  8.75    750 400 1000
#> 16      55 127 10.50    750 400 1000
#> 
#> $shrubData
#>   Species Cover Height Z50  Z95
#> 1      65     2     30 300  800
#> 2      61     3    100 300  500
#> 3      52     2    120 300  800
#> 4      78     3    100 300 1360
#> 5      88     2    150 300  800
#> 6      39     3     50 300  800
#> 7      46     3     30 300  800
#> 8      79    10    150 300  800
#> 9      31     2     30 300  800
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"  

soilpar = defaultSoilParams()
f2 = estimateRootingDepth(f1, soilpar, 300, 100, SpParamsMED)
print(f2)

#> $ID
#> [1] "81065"
#> 
#> $patchsize
#> [1] 10000
#> 
#> $treeData
#>    Species   N   DBH Height     Z50 Z95
#> 1       55 127  9.55    750 220.251 500
#> 2       55 127  9.40    650 220.251 500
#> 3       55  32 21.00   1600 220.251 500
#> 4       55  14 23.30   1450 220.251 500
#> 5       55  32 22.15   1450 220.251 500
#> 6       55  14 23.10   1450 220.251 500
#> 7       55  32 19.00   1350 220.251 500
#> 8       55  14 25.40   1450 220.251 500
#> 9       55  32 18.50   1250 220.251 500
#> 10      55 127  8.20    750 220.251 500
#> 11      55  32 21.25   1400 220.251 500
#> 12      55  32 19.30   1300 220.251 500
#> 13      55  14 33.95   1900 220.251 500
#> 14      55 127 11.35    750 220.251 500
#> 15      55 127  8.75    750 220.251 500
#> 16      55 127 10.50    750 220.251 500
#> 
#> $shrubData
#>   Species Cover Height      Z50 Z95
#> 1      65     2     30 250.0000 500
#> 2      61     3    100 250.0000 500
#> 3      52     2    120 250.0000 500
#> 4      78     3    100 250.0000 500
#> 5      88     2    150 250.0000 500
#> 6      39     3     50       NA  NA
#> 7      46     3     30       NA  NA
#> 8      79    10    150 197.5605 500
#> 9      31     2     30       NA  NA
#> 
#> $herbCover
#> [1] NA
#> 
#> $herbHeight
#> [1] NA
#> 
#> attr(,"class")
#> [1] "forest" "list"  

l1 = IFN2forestlist(exampletreedata, exampleshrubdata, SpParamsMED)

#> Number of plots: 5
#> Filtered records in tree data: 20
#> Translating species codes...
#> Shrub data records with unrecognized IFN species codes: 4/43
#> 
#> 130 141 
#>   1   3 
#> Extracting IFN data...
#> done.

l2 = estimateRootingDepthList(l1, soilpar, 300, 100, SpParamsMED)