Estimate Rooting Depth Parameters

Estimates rooting depth parameters for forest objects following Cabon et al. (2018).

estimateRootingDepth(
  forest,
  soil,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

estimateRootingDepthList(
  forestlist,
  soillist,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

Arguments

forest

An object of class forest.

soil

An object of class soil or a data frame with soil parameters.

PET_summer

Average (climatic) potential evapotranspiration (mm) corresponding to summer months.

P_summer

Average (climatic) precipitation (mm) corresponding to summer months.

SpParams

A data frame with species parameters (see SpParamsMED).

fillMissing

A boolean flag to indicate that missing estimates should be filled with mean values.

forestlist

A list of objects of class forest.

soillist

A list with objects of class soil or data frames with soil parameters.

Value

Function estimateRootingDepth returns object of class forest with modified values for Z50 and Z95. Function estimateRootingDepthList returns a list of forest objects with modified root distribution parameters.

Details

If PET_summer or P_summer are of length one, their elements are repeated for all elements in forestlist. Similarly, if soillist is an object of class soil or a data frame with soil parameters, the same soil parameters are repeated for all elements in forestlist.

References

Cabon, A., Martinez-Vilalta, J., de Aragon, J.M., Poyatos, R., De Caceres, M., 2018. Applying the eco-hydrological equilibrium hypothesis to model root distribution in water-limited forests. Ecohydrology 11, e2015. https://doi.org/10.1002/eco.2015

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

# Builds from IFN3 data a list whose elements are 'forest' objects
data(IFN_species_mapping)
#> Warning: data set ‘IFN_species_mapping’ not found
data(piesMayoresIFN3)
data(regeneraIFN3)
data(matorralIFN3)
l1 <- IFN2forest(piesMayoresIFN3, IFN_species_mapping, SpParamsMED,
               matorral = matorralIFN3, regenera = regeneraIFN3)
#> Number of plots: 5
#> Filtering missing and zero values...
#> Filtering small trees (DBH < 1) ...
#> Filtering dead trees...
#> Translating species codes...
#> Error in translateSpeciesCodes(x = x$Especie, species_names = IFN_species_mapping$Name,     species_codes = IFN_species_mapping$Codes): could not find function "translateSpeciesCodes"

# Estimate rooting depth
soilpar <- defaultSoilParams()
l2 <- estimateRootingDepthList(l1, soilpar, 300, 100, SpParamsMED)
#> Error in eval(expr, envir, enclos): object 'l1' not found

# Compare tree rooting depths
print(l1[[1]]$treeData)
#> Error in eval(expr, envir, enclos): object 'l1' not found
print(l2[[1]]$treeData)
#> Error in eval(expr, envir, enclos): object 'l2' not found