Estimate Rooting Depth Parameters

Estimates rooting depth parameters for forest objects following Cabon et al. (2018).

estimateRootingDepth(
  forest,
  soil,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

estimateRootingDepthList(
  forestlist,
  soillist,
  PET_summer,
  P_summer,
  SpParams,
  fillMissing = FALSE
)

Arguments

forest

An object of class forest.

soil

An object of class soil or a data frame with soil parameters.

PET_summer

Average (climatic) potential evapotranspiration (mm) corresponding to summer months.

P_summer

Average (climatic) precipitation (mm) corresponding to summer months.

SpParams

A data frame with species parameters (see SpParamsMED).

fillMissing

A boolean flag to indicate that missing estimates should be filled with mean values.

forestlist

A list of objects of class forest.

soillist

A list with objects of class soil or data frames with soil parameters.

Value

Function estimateRootingDepth returns object of class forest with modified values for Z50 and Z95. Function estimateRootingDepthList returns a list of forest objects with modified root distribution parameters.

Details

If PET_summer or P_summer are of length one, their elements are repeated for all elements in forestlist. Similarly, if soillist is an object of class soil or a data frame with soil parameters, the same soil parameters are repeated for all elements in forestlist.

References

Cabon, A., Martinez-Vilalta, J., de Aragon, J.M., Poyatos, R., De Caceres, M., 2018. Applying the eco-hydrological equilibrium hypothesis to model root distribution in water-limited forests. Ecohydrology 11, e2015. https://doi.org/10.1002/eco.2015

See also

Author

Miquel De Cáceres Ainsa, EMF-CREAF

Examples

# Builds from IFN3 data a list whose elements are 'forest' objects
data(IFN_species_mapping)
data(piesMayoresIFN3)
data(regeneraIFN3)
data(matorralIFN3)
l1 <- IFN2forest(piesMayoresIFN3, IFN_species_mapping, SpParamsMED,
               matorral = matorralIFN3, regenera = regeneraIFN3)
#> Number of plots: 5
#> Filtering missing and zero values...
#> Filtering small trees (DBH < 1) ...
#> Filtering dead trees...
#> Translating species codes...
#> Filtering wrong growth forms ...
#> Setting default root distribution ...
#> Building forest objects...
#> done.

# Estimate rooting depth
soilpar <- defaultSoilParams()
l2 <- estimateRootingDepthList(l1, soilpar, 300, 100, SpParamsMED)

# Compare tree rooting depths
print(l1[[1]]$treeData)
#>        Species  DBH Height           N      Z50 Z95 OrdenIf2 OrdenIf3
#> 1  Pinus nigra 19.1    750   31.830989 173.1933 813      000      001
#> 2  Pinus nigra 18.8    650   31.830989 173.1933 813      000      002
#> 3  Pinus nigra 42.0   1600   14.147106 173.1933 813      002      003
#> 4  Pinus nigra 46.6   1450    5.092958 173.1933 813      000      004
#> 5  Pinus nigra 44.3   1450    5.092958 173.1933 813      004      005
#> 6  Pinus nigra 46.2   1450    5.092958 173.1933 813      000      006
#> 7  Pinus nigra 38.0   1350   14.147106 173.1933 813      005      007
#> 8  Pinus nigra 50.8   1450    5.092958 173.1933 813      009      008
#> 9  Pinus nigra 37.0   1250   14.147106 173.1933 813      011      009
#> 10 Pinus nigra 16.4    750   31.830989 173.1933 813      000      010
#> 11 Pinus nigra 42.5   1400    5.092958 173.1933 813      013      011
#> 12 Pinus nigra 38.6   1300   14.147106 173.1933 813      014      012
#> 13 Pinus nigra 67.9   1900    5.092958 173.1933 813      017      013
#> 14 Pinus nigra 22.7    750   14.147106 173.1933 813      000      014
#> 15 Pinus nigra 17.5    750   31.830989 173.1933 813      000      015
#> 16 Pinus nigra 21.0    750   31.830989 173.1933 813      000      016
#> 17 Pinus nigra  5.0    350 1400.563501 173.1933 813     <NA>     <NA>
#> 18 Pinus nigra  1.5    100 1273.239546 173.1933 813     <NA>     <NA>
print(l2[[1]]$treeData)
#>        Species  DBH Height           N      Z50 Z95 OrdenIf2 OrdenIf3
#> 1  Pinus nigra 19.1    750   31.830989 207.6492 500      000      001
#> 2  Pinus nigra 18.8    650   31.830989 207.6492 500      000      002
#> 3  Pinus nigra 42.0   1600   14.147106 207.6492 500      002      003
#> 4  Pinus nigra 46.6   1450    5.092958 207.6492 500      000      004
#> 5  Pinus nigra 44.3   1450    5.092958 207.6492 500      004      005
#> 6  Pinus nigra 46.2   1450    5.092958 207.6492 500      000      006
#> 7  Pinus nigra 38.0   1350   14.147106 207.6492 500      005      007
#> 8  Pinus nigra 50.8   1450    5.092958 207.6492 500      009      008
#> 9  Pinus nigra 37.0   1250   14.147106 207.6492 500      011      009
#> 10 Pinus nigra 16.4    750   31.830989 207.6492 500      000      010
#> 11 Pinus nigra 42.5   1400    5.092958 207.6492 500      013      011
#> 12 Pinus nigra 38.6   1300   14.147106 207.6492 500      014      012
#> 13 Pinus nigra 67.9   1900    5.092958 207.6492 500      017      013
#> 14 Pinus nigra 22.7    750   14.147106 207.6492 500      000      014
#> 15 Pinus nigra 17.5    750   31.830989 207.6492 500      000      015
#> 16 Pinus nigra 21.0    750   31.830989 207.6492 500      000      016
#> 17 Pinus nigra  5.0    350 1400.563501 207.6492 500     <NA>     <NA>
#> 18 Pinus nigra  1.5    100 1273.239546 207.6492 500     <NA>     <NA>