Optimization of rock fragment content

Function spwb_rockOptimization finds optimum rock fragment content in the soil corresponding to given vegetation, weather and target percent loss of conductance (PLC), following the method described in Druel et al. (2023).

spwb_rockOptimization(
  x,
  meteo,
  PLC90_target = 12,
  PLC90_tol = 0.5,
  max_simu = 7,
  model_varLim = 10,
  max_rocks = 99,
  verbose = FALSE,
  ...
)

Arguments

x

An object of class spwbInput.

meteo

A data frame with daily meteorological data series (see spwb).

PLC90_target

Stem PLC target (quantile 90).

PLC90_tol

Limit of the PLC target tolerance, only in some conditions.

max_simu

Maximum of simulation authorized before stopping.

model_varLim

Limit of the soil extractable water (SEW) variation of the model accepted.

max_rocks

Maximum content in coarse fragments allowed for any soil layer.

verbose

A logical value. Print the internal messages of the function?

...

Additional parameters to function spwb.

Value

Function spwb_rockOptimization returns a list containing the estimated rock fragment content, the corresponding soil extractable water and a dataframe of simulation results needed for the estimation.

Details

The function performs a model inversion based on an ecohydrological assumption, consisting in that forest leaf area index is in equilibrium with a low embolism rate under normal conditions. This is translated in that the 90% interannual quantile of the maximum annual percent loss of conductance (PLC), averaged over plant cohorts, should be close to a target PLC value (by default 12%).

References

Druel, A., Martins, N., Cochard, H., De Caceres, M., Delzon, S., Mencuccini, M., Torres-Ruiz, J., and Ruffault, J.: European forest vulnerability to hydraulic failure: an ecohydrological approach, EGU General Assembly 2023, Vienna, Austria, 24–28 Apr 2023, EGU23-17068, https://doi.org/10.5194/egusphere-egu23-17068, 2023.

See also

Author

Arsène Druel, URFM-INRAE

Nicolas Martin-StPaul, URFM-INRAE

Miquel De Cáceres Ainsa, CREAF

Examples

# \donttest{
#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforestMED)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with default soil params
examplesoil <- soil(defaultSoilParams(2))

#Initialize control parameters
control <- defaultControl("Granier")

#Initialize input
x <- forest2spwbInput(exampleforestMED,examplesoil, SpParamsMED, control)

#Rock fragment content optimization
spwb_rockOptimization(x, meteo = examplemeteo,
                      elevation = 100, latitude = 41.82592)
#> $SEW
#> [1] 30.69065
#> 
#> $RFC
#> [1] 53.32618 95.98713
#> 
#> $SimulationTable
#>      SEW       PLC90
#> 1   0.00 100.0000000
#> 2 195.02   0.0000000
#> 3  50.00   0.4335227
#> 4   1.95  98.9754471
#> 5  25.98  24.7050476
#> 
# }